requirements.txt
Flask==3.0.0
onnxruntime==1.17.1
Pillow==10.2.0
numpy==1.26.4
waitress==2.1.2
